requirements.txt
ansible-core==2.14.3
ansible-lint==5.4.0 # lock
black==23.1.0
dotty-dict==1.3.1
GitPython==3.1.30 # Temporarily lock this due to a version conflict
graphviz==0.20.1
isort==5.12.0
luadata==1.0.5
lupa==1.14.1
mypy==1.0.1
PyDriller==2.4.1
pylint==2.16.3
PyYAML==6.0
requests==2.28.2
semver==2.13.0
types-PyYAML==6.0.12.8
types-requests==2.28.11.15
types-urllib3==1.26.25.8
yamlfmt==1.1.1
yamllint==1.29.0
## The following requirements were added by pip freeze:
astroid==2.14.2
bracex==2.3.post1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.2
dill==0.3.6
enrich==1.2.7
gitdb==4.0.10
idna==3.4
Jinja2==3.1.2
lazy-object-proxy==1.9.0
lizard==1.17.10
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycparser==2.21
Pygments==2.14.0
pytz==2022.7.1
resolvelib==0.8.1
rich==13.3.1
ruamel.yaml.clib==0.2.7
ruamel.yaml==0.17.21
smmap==5.0.0
tenacity==8.2.2
tomli==2.0.1
tomlkit==0.11.6
types-pytz==2022.7.1.2
typing_extensions==4.5.0
urllib3==1.26.14
wcmatch==8.4.1
wrapt==1.15.0
